refactor(subtitles): extract RTL text punctuation fix into dedicated PlayerSubtitleRtlFix helper
