+++ app/src/main/java/com/nuvio/tv/ui/screens/player/PlayerRuntimeControllerInitialization.kt
@@ -2256,7 +2256,7 @@ private class CueNormalizingTextOutput(
 
     private fun processCue(cue: Cue): Cue {
         var processed = stripReplacementCharacter(cue)
-        processed = fixRtlCueText(processed)
+        processed = PlayerSubtitleRtlFix.fixCueText(processed, isBuiltInSubtitleProvider())
         if (shouldNormalizeCuePositionProvider()) {
             processed = normalizeCuePosition(processed)
         }
@@ -2314,271 +2314,6 @@ private class CueNormalizingTextOutput(
             }
         }
         return cue.buildUpon().setText(builder).build()
-    }
-
-    private fun fixRtlCueText(cue: Cue): Cue {
-        val text = cue.text ?: return cue
-        if (!hasAnyRtlCharacter(text)) {
-            return cue
-        }
-
-        // Arabic: wrap each physical line with RLE (\u202B) ... PDF (\u202C).
-        // This renders boundary punctuation and auto-wrapped lines as RTL in an LTR container.
-        if (containsArabic(text)) {
-            val builder = android.text.SpannableStringBuilder()
-            val lines = text.splitByNewlines()
-            for (i in lines.indices) {
-                if (i > 0) builder.append("\n")
-                // Clear existing directional markers -> prevents double wrapping upon re-execution (idempotent).
-                val line = lines[i].stripDirectionalWrap()
-                if (line.isEmpty()) {
-                    builder.append(line)
-                    continue
-                }
-                // Keep the trailing CR (paragraph separator) OUTSIDE of the embedding; otherwise
-                // it terminates the RLE run and leaves the PDF orphan.
-                val hasCr = line[line.length - 1] == '\r'
-                val core = if (hasCr) line.subSequence(0, line.length - 1) else line
-                if (core.isEmpty()) {
-                    builder.append(line)
-                    continue
-                }
-                builder.append("\u202B").append(core).append("\u202C")
-                if (hasCr) builder.append("\r")
-            }
-            if (builder.contentEquals(text)) return cue
-            return cue.buildUpon().setText(builder).build()
-        }
-
-        // Hebrew / other RTL: punctuation boundary-swap method (span preserving).
-        if (containsRtlChars(text)) {
-            val isBuiltIn = isBuiltInSubtitleProvider()
-            val builder = android.text.SpannableStringBuilder()
-            val lines = text.splitByNewlines()
-            var changed = false
-            for (i in lines.indices) {
-                if (i > 0) builder.append("\n")
-                val line = lines[i]
-                val fixed = if (isBuiltIn) {
-                    moveLeadingRtlPunctuationToEndForBuiltIn(line)
-                } else {
-                    fixRtlPunctuationForLtr(line)
-                }
-                if (fixed !== line) changed = true
-                builder.append(fixed)
-            }
-            if (!changed) return cue
-            return cue.buildUpon().setText(builder).build()
-        }
-
-        return cue
-    }
-
-    private fun containsArabic(text: CharSequence): Boolean {
-        var i = 0
-        while (i < text.length) {
-            val codePoint = Character.codePointAt(text, i)
-            if (codePoint in 0x0600..0x06FF || // Arabic block
-                codePoint in 0x0750..0x077F || // Arabic Supplement
-                codePoint in 0x0870..0x08FF || // Arabic Extended
-                codePoint in 0xFB50..0xFDFF || // Arabic Presentation Forms-A
-                codePoint in 0xFE70..0xFEFF || // Arabic Presentation Forms-B
-                Character.getDirectionality(codePoint) == Character.DIRECTIONALITY_RIGHT_TO_LEFT_ARABIC
-            ) {
-                return true
-            }
-            i += Character.charCount(codePoint)
-        }
-        return false
-    }
-
-    private fun mirrorPunctuation(c: Char): Char = when (c) {
-        '(' -> ')'
-        ')' -> '('
-        else -> c
-    }
-
-    private fun appendMirroredReversed(
-        out: android.text.SpannableStringBuilder,
-        line: CharSequence,
-        from: Int,
-        toExclusive: Int
-    ) {
-        if (from >= toExclusive) return
-    
-        fun isNumberSeparator(c: Char) = c == ',' || c == ':' || c == '.' || c == '-'
-    
-        // 1. Split [from, toExclusive) into chunks: number-runs (digits + embedded , : .) stay together
-        val chunks = ArrayList<IntRange>()
-        var i = from
-        while (i < toExclusive) {
-            if (line[i].isDigit()) {
-                val start = i
-                i++
-                while (i < toExclusive) {
-                    if (line[i].isDigit()) {
-                        i++
-                    } else if (
-                        isNumberSeparator(line[i]) &&
-                        i + 1 < toExclusive &&
-                        line[i + 1].isDigit()
-                    ) {
-                        // separator sandwiched between digits, e.g. 16,300 / 10:50 / 1.23
-                        i++ // consume separator, loop will consume following digits
-                    } else {
-                        break
-                    }
-                }
-                chunks.add(start until i)          // whole number (with separators) as one chunk
-            } else {
-                chunks.add(i until i + 1)           // single char chunk
-                i++
-            }
-        }
-    
-        // 2. Walk chunks back-to-front, but append each chunk's *contents* in original order
-        for (idx in chunks.indices.reversed()) {
-            val range = chunks[idx]
-            if (range.last - range.first + 1 > 1) {
-                // number run -> keep as-is, don't reverse the digits/separators themselves
-                out.append(line.subSequence(range.first, range.last + 1))
-            } else {
-                val c = line[range.first]
-                val m = mirrorPunctuation(c)
-                out.append(if (m != c) m.toString() else line.subSequence(range.first, range.first + 1))
-            }
-        }
-    }
-    
-    // Take CharSequence instead of String -> preserve spans.
-    private fun fixRtlPunctuationForLtr(line: CharSequence): CharSequence {
-        if (line.isEmpty()) return line
-        val hasCr = line[line.length - 1] == '\r'
-        val end0 = if (hasCr) line.length - 1 else line.length
-        if (end0 == 0) return line
-
-        var start = 0
-        while (start < end0 && isRtlPunctuation(line[start], isEnd = false)) start++
-
-        var end = end0
-        while (end > start && isRtlPunctuation(line[end - 1], isEnd = true)) end--
-
-        if (start == 0 && end == end0) return line
-
-        val out = android.text.SpannableStringBuilder()
-        appendMirroredReversed(out, line, end, end0)   // trailing punct/numbers -> front
-        out.append(line.subSequence(start, end))       // middle, untouched
-        appendMirroredReversed(out, line, 0, start)    // leading punct/numbers -> end
-        if (hasCr) out.append("\r")
-        return out
-    }
-
-    private fun moveLeadingRtlPunctuationToEndForBuiltIn(line: CharSequence): CharSequence {
-        if (line.isEmpty()) return line
-        val hasCr = line[line.length - 1] == '\r'
-        val end0 = if (hasCr) line.length - 1 else line.length
-        if (end0 == 0) return line
-
-        var end = 0
-        while (end < end0 && line[end] in MOBILE_RTL_PUNCTUATION) end++
-        if (end == 0) return line
-
-        val out = android.text.SpannableStringBuilder()
-        out.append(line.subSequence(end, end0))
-            .append(line.subSequence(0, end))
-        if (hasCr) out.append("\r")
-        return out
-    }
-
-    // Clears existing directional control characters (idempotency + legacy RLM/LRE remnants).
-    private fun CharSequence.stripDirectionalWrap(): CharSequence {
-        val hasMarker = (0 until length).any { isDirectionalMark(this[it]) }
-        if (!hasMarker) return this
-        val sb = android.text.SpannableStringBuilder(this)
-        var k = 0
-        while (k < sb.length) {
-            if (isDirectionalMark(sb[k])) sb.delete(k, k + 1) else k++
-        }
-        return sb
-    }
-
-    private fun isDirectionalMark(c: Char): Boolean =
-        c == '\u202A' || c == '\u202B' || c == '\u202C' || // LRE / RLE / PDF
-        c == '\u200E' || c == '\u200F'                     // LRM / RLM
-
-    private fun CharSequence.splitByNewlines(): List<CharSequence> {
-        val result = mutableListOf<CharSequence>()
-        var start = 0
-        var i = 0
-        while (i < this.length) {
-            if (this[i] == '\n') {
-                result.add(this.subSequence(start, i))
-                start = i + 1
-            }
-            i++
-        }
-        result.add(this.subSequence(start, this.length))
-        return result
-    }
-
-    private fun isRtlPunctuation(ch: Char, isEnd: Boolean): Boolean {
-        if (isEnd && ch.isDigit()) return false
-        return ch in RTL_PUNCTUATION || ch.isWhitespace()
-    }
-
-    private fun containsRtlChars(text: CharSequence): Boolean {
-        var i = 0
-        while (i < text.length) {
-            val codePoint = Character.codePointAt(text, i)
-            
-            // Direct Unicode range checks for Hebrew and Arabic scripts
-            if (codePoint in 0x0590..0x05FF || // Hebrew block (letters, points, punctuation)
-                codePoint in 0xFB1D..0xFB4F || // Hebrew Presentation Forms
-                codePoint in 0x0600..0x06FF || // Arabic block
-                codePoint in 0x0750..0x077F || // Arabic Supplement
-                codePoint in 0x0870..0x08FF || // Arabic Extended
-                codePoint in 0xFB50..0xFDFF || // Arabic Presentation Forms-A
-                codePoint in 0xFE70..0xFEFF    // Arabic Presentation Forms-B
-            ) {
-                return true
-            }
-            
-            val d = Character.getDirectionality(codePoint)
-            if (d == Character.DIRECTIONALITY_RIGHT_TO_LEFT ||
-                d == Character.DIRECTIONALITY_RIGHT_TO_LEFT_ARABIC ||
-                d == Character.DIRECTIONALITY_ARABIC_NUMBER) return true
-            i += Character.charCount(codePoint)
-        }
-        return false
-    }
-
-    private fun hasAnyRtlCharacter(text: CharSequence): Boolean {
-        var i = 0
-        val len = text.length
-        while (i < len) {
-            val codePoint = Character.codePointAt(text, i)
-            if (codePoint >= 0x0590) {
-                if (codePoint in 0x0590..0x08FF ||
-                    codePoint in 0xFB1D..0xFEFF
-                ) {
-                    return true
-                }
-                val d = Character.getDirectionality(codePoint)
-                if (d == Character.DIRECTIONALITY_RIGHT_TO_LEFT ||
-                    d == Character.DIRECTIONALITY_RIGHT_TO_LEFT_ARABIC ||
-                    d == Character.DIRECTIONALITY_ARABIC_NUMBER
-                ) {
-                    return true
-                }
-            }
-            i += Character.charCount(codePoint)
-        }
-        return false
-    }
-
-    companion object {
-        private val RTL_PUNCTUATION = setOf('.', ',', '?', '!', '-', ':', ';', '…', ')', '(', '\'', '"') + ('0'..'9')
-        private val MOBILE_RTL_PUNCTUATION = setOf('.', ',', '?', '!', '-', ':', ';', '…', ')', '(')
     }
 }
